Address second round of Copilot code review comments
Refactorings and improvements:
- Extract duplicate PaperTrail helper methods to shared concern
- Use version ID ordering instead of created_at for timestamp precision
- Replace brittle JavaScript URL construction with Rails path helpers
- Update outdated comment about batch processing strategy

Changes:
- Created PaperTrailHelpers concern for safe deserialization and markdown detection
- Updated ManifestationController and AdminController to include the shared concern
- Modified version_diff to use version ID instead of created_at for next version lookup
- Refactored versions.html.haml to use data-diff-url attribute with Rails helper
- Simplified JavaScript to read pre-built URL from data attribute
- Fixed comment mismatch in duplicate_works method
