+++ app/controllers/manifestation_controller.rb
@@ -7,6 +7,7 @@ class ManifestationController < ApplicationController
   include Tracking
   include BybeUtils
   include KwicConcordanceConcern
+  include PaperTrailHelpers
 
   before_action only: %i(list show remove_link edit_metadata add_aboutnesses versions version_diff
                          restore_version) do |c|
@@ -225,8 +226,9 @@ class ManifestationController < ApplicationController
       @previous_markdown = ''
     end
 
-    # Get the version after this change by finding the next version
-    next_version = @m.versions.where('created_at > ?', version.created_at).order(created_at: :asc).first
+    # Get the version after this change by finding the next version by ID
+    # Using ID is more reliable than created_at (avoids timestamp precision issues)
+    next_version = @m.versions.where('id > ?', version.id).order(id: :asc).first
     if next_version && next_version.object.present?
       new_attrs = deserialize_version_object(next_version.object)
       @current_markdown = new_attrs['markdown'] || ''
@@ -266,41 +268,6 @@ class ManifestationController < ApplicationController
   end
 
   private
-
-  # Safe deserialization of PaperTrail version objects
-  def deserialize_version_object(object_yaml)
-    # Use safe_load with allowed classes for ActiveRecord objects
-    Psych.safe_load(
-      object_yaml,
-      permitted_classes: [
-        ActiveSupport::TimeWithZone,
-        ActiveSupport::TimeZone,
-        Time,
-        Symbol
-      ],
-      aliases: true
-    )
-  end
-
-  # Check if markdown field was changed in a version
-  def markdown_changed_in_version?(version)
-    # If object_changes is available, check if markdown key exists
-    return version.changeset&.key?('markdown') || false if version.object_changes.present?
-
-    # Fallback: Check by comparing deserialized objects
-    # First version (create event) always has markdown
-    return true if version.event == 'create'
-
-    # For updates without object_changes, check if object contains markdown
-    return false if version.object.blank?
-
-    begin
-      old_attrs = deserialize_version_object(version.object)
-      old_attrs.key?('markdown')
-    rescue StandardError
-      false
-    end
-  end
 
   def fetch_new_authors(since)
     # Authorities created recently with published manifestations
